feat: migrate landing page to appdir
