+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "typescript.tsdk": "node_modules/.pnpm/typescript@4.7.3/node_modules/typescript/lib",
+  "typescript.enablePromptUseWorkspaceTsdk": true
+}
