+++ package.json
@@ -37,7 +37,7 @@
     "fuse.js": "^6.6.2",
     "legacy-modes": "npm:@codemirror/legacy-modes@^6.3.1",
     "next": "^13.3.0",
-    "next-auth": "^4.10.3",
+    "next-auth": "^4.22.0",
     "next-mdx-remote": "^4.1.0",
     "next-themes": "^0.2.0",
     "prism-react-renderer": "^1.3.5",
