+++ src/app/providers.tsx
@@ -1,0 +1,8 @@
+'use client'
+import React from 'react'
+
+import { ThemeProvider } from 'next-themes'
+
+export function Providers({ children }: { children: React.ReactNode }) {
+  return <ThemeProvider>{children}</ThemeProvider>
+}
