+++ src/components/Landing/IncrementalNumber.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useRef } from 'react'
 
 import CountUp from 'react-countup'
