+++ src/pages/api/auth/[...nextauth].ts
@@ -1,64 +1,5 @@
-import { PrismaAdapter } from '@next-auth/prisma-adapter'
 import NextAuth from 'next-auth'
-import type { NextAuthOptions } from 'next-auth'
-import GithubProvider from 'next-auth/providers/github'
-import GoogleProvider from 'next-auth/providers/google'
 
-import prisma from '@/lib/prisma'
-
-if (
-  !process.env.GITHUB_ID ||
-  !process.env.GITHUB_SECRET ||
-  !process.env.GOOGLE_CLIENT_ID ||
-  !process.env.GOOGLE_CLIENT_SECRET
-)
-  throw new Error('Failed to initialize authentication')
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID,
-      clientSecret: process.env.GITHUB_SECRET,
-      profile(profile) {
-        return {
-          id: profile.id.toString(),
-          name: profile.name || profile.login,
-          username: profile.login,
-          email: profile.email,
-          image: profile.avatar_url
-        }
-      }
-    }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-      profile(profile) {
-        return {
-          id: profile.sub,
-          name: profile.name,
-          username: profile.name,
-          email: profile.email,
-          image: profile.picture
-        }
-      }
-    })
-  ],
-  pages: {
-    signIn: '/login',
-    error: '/login'
-  },
-  callbacks: {
-    session: ({ session, user }) => ({
-      ...session,
-      user: {
-        ...session.user,
-        id: user.id,
-        username: user.username,
-        admin: user.admin
-      }
-    })
-  }
-}
+import { authOptions } from '@/lib/auth'
 
 export default NextAuth(authOptions)
